# src/main/java/ru/skypro/homework/dto/AdsDto.java
package ru.skypro.homework.dto;

import lombok.Data;

@Data
public class AdsDto {
    private String image;
    private Integer author;
    private Integer price;
    private Integer pk;
    private String title;
}